fix: organize python api integration test folder structur
