+++ integration_test/question/get_question_detail_test.py
@@ -1,4 +1,4 @@
-from utils import utils
+from integration_test import utils
 
 
 def test_error_question_not_found():
