+++ integration_test/utils/__init__.py
@@ -1,0 +1,4 @@
+from integration_test.utils.answer import *
+from integration_test.utils.question import *
+from integration_test.utils.user import *
+from integration_test.utils.vote_option import *
