+++ integration_test/utils/vote_option.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+
+class VoteOption(Enum):
+    ThumbsUp = "thumbsup"
+    ThumbsDown = "thumbsdown"
